Add brute force password attack implementation
The code adds a new file `ataques/forcaBruta.js` that contains the implementation of a brute force password attack. It imports the `createHash` function from the `crypto` module to hash passwords using the SHA-512 algorithm. The code defines a `Usuario` class with a constructor that takes a name and a password, and a `autentica` method to check if a given name and password match the stored name and hash. A new instance of `Usuario` is created with the name "matheus" and password "1337". Then, a loop tests password combinations from 0 to 9999 and calls the `autentica` method with each combination. If a match is found, it outputs the discovered password; otherwise, it outputs the tested password combination.