feat: Add Usuario class and authentication methods
This commit adds a new file "dicionario.js" which defines a class called Usuario. The class has a constructor that takes a nome (name) and senha (password), and creates a hash using the sha256 algorithm. It also has a method to authenticate the user by comparing the nome and senha with the stored values.

The commit also adds a new file "rainbowTable.js" which defines a function to create a hash using a given strategy. It creates a rainbow table of common passwords along with their corresponding hashes. It then checks if any of the leaked hashes match the hashes in the rainbow table and logs the corresponding password if a match is found.